Move EthTxSigs.signMessage(hash,privKey) to new util class in test-clients
