+++ hedera-node/hapi-utils/src/main/java/com/hedera/node/app/hapi/utils/ethereum/EthTxSigs.java
@@ -27,9 +27,7 @@ import com.esaulpaugh.headlong.rlp.RLPEncoder;
 import com.esaulpaugh.headlong.util.Integers;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.MoreObjects;
-import com.sun.jna.ptr.IntByReference;
 import com.sun.jna.ptr.LongByReference;
-import java.math.BigInteger;
 import java.nio.ByteBuffer;
 import java.util.Arrays;
 import org.apache.commons.codec.binary.Hex;
@@ -45,76 +43,6 @@ public record EthTxSigs(byte[] publicKey, byte[] address) {
         byte[] compressedKey = recoverCompressedPubKey(pubKey);
 
         return new EthTxSigs(compressedKey, address);
-    }
-
-    public static EthTxData signMessage(EthTxData ethTx, byte[] privateKey) {
-        byte[] signableMessage = calculateSignableMessage(ethTx);
-        final LibSecp256k1.secp256k1_ecdsa_recoverable_signature signature =
-                new LibSecp256k1.secp256k1_ecdsa_recoverable_signature();
-        LibSecp256k1.secp256k1_ecdsa_sign_recoverable(
-                CONTEXT, signature, new Keccak.Digest256().digest(signableMessage), privateKey, null, null);
-
-        final ByteBuffer compactSig = ByteBuffer.allocate(64);
-        final IntByReference recId = new IntByReference(0);
-        LibSecp256k1.secp256k1_ecdsa_recoverable_signature_serialize_compact(
-                LibSecp256k1.CONTEXT, compactSig, recId, signature);
-        compactSig.flip();
-        final byte[] sig = compactSig.array();
-
-        // wrap in signature object
-        final byte[] r = new byte[32];
-        System.arraycopy(sig, 0, r, 0, 32);
-        final byte[] s = new byte[32];
-        System.arraycopy(sig, 32, s, 0, 32);
-
-        BigInteger val;
-        // calulations originate from https://eips.ethereum.org/EIPS/eip-155
-        if (ethTx.type() == LEGACY_ETHEREUM) {
-            if (ethTx.chainId() == null || ethTx.chainId().length == 0) {
-                val = BigInteger.valueOf(27L + recId.getValue());
-            } else {
-                val = BigInteger.valueOf(35L + recId.getValue())
-                        .add(new BigInteger(1, ethTx.chainId()).multiply(BigInteger.TWO));
-            }
-        } else {
-            val = null;
-        }
-
-        return new EthTxData(
-                ethTx.rawTx(),
-                ethTx.type(),
-                ethTx.chainId(),
-                ethTx.nonce(),
-                ethTx.gasPrice(),
-                ethTx.maxPriorityGas(),
-                ethTx.maxGas(),
-                ethTx.gasLimit(),
-                ethTx.to(),
-                ethTx.value(),
-                ethTx.callData(),
-                ethTx.accessList(),
-                (byte) recId.getValue(),
-                val == null ? null : val.toByteArray(),
-                r,
-                s);
-    }
-
-    public static byte[] signMessage(final byte[] messageHash, byte[] privateKey) {
-        final LibSecp256k1.secp256k1_ecdsa_recoverable_signature signature =
-                new LibSecp256k1.secp256k1_ecdsa_recoverable_signature();
-        LibSecp256k1.secp256k1_ecdsa_sign_recoverable(CONTEXT, signature, messageHash, privateKey, null, null);
-
-        final ByteBuffer compactSig = ByteBuffer.allocate(64);
-        final IntByReference recId = new IntByReference(0);
-        LibSecp256k1.secp256k1_ecdsa_recoverable_signature_serialize_compact(
-                LibSecp256k1.CONTEXT, compactSig, recId, signature);
-        compactSig.flip();
-        final byte[] sig = compactSig.array();
-
-        final byte[] result = new byte[65];
-        System.arraycopy(sig, 0, result, 0, 64);
-        result[64] = (byte) (recId.getValue() + 27);
-        return result;
     }
 
     public static byte[] calculateSignableMessage(EthTxData ethTx) {
