+++ hedera-node/test-clients/src/main/java/module-info.java
@@ -74,6 +74,7 @@ module com.hedera.node.test.clients {
     requires com.fasterxml.jackson.databind;
     requires com.github.docker.java.api;
     requires com.hedera.evm;
+    requires com.sun.jna;
     requires io.grpc.netty;
     requires io.grpc.stub;
     requires io.netty.handler;
@@ -84,6 +85,7 @@ module com.hedera.node.test.clients {
     requires org.bouncycastle.provider;
     requires org.hyperledger.besu.datatypes;
     requires org.hyperledger.besu.internal.crypto;
+    requires org.hyperledger.besu.nativelib.secp256k1;
     requires org.json;
     requires org.junit.platform.commons;
     requires org.opentest4j;
